portal/backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
bcrypt==3.2.2
boto3==1.20.44
botocore==1.23.54
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-extensions==3.0.9
django-import-export==2.7.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-security==0.12.0
django-sslserver==0.22
django-timezone-field==4.0
django-webpack-loader==0.7.0
djangorestframework==3.11.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
jenkins==1.0.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupPy==1.14
MarkupSafe==2.1.1
multi-key-dict==2.0.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
paramiko==2.11.0
pbr==5.9.0
protobuf==4.21.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.0
python-jenkins==1.7.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
jsonschema==4.8.0
